+++ templates/changeItemInstanceDescription/test.d
@@ -1,9 +1,8 @@
 /*
  * #{ISSUE_NUM} {LONGNAME}
  */
-func int G1CP_Test_{ISSUE_NUM_PAD}() {
+func void G1CP_Test_{ISSUE_NUM_PAD}() {
     {LANGCHECK}
     var C_Item itm; itm = G1CP_Testsuite_CreateItem("### TODO Item instance name ###");
-
-    return G1CP_Testsuite_InspectItemString(itm, "description", "### TODO Expected string ###");
+    G1CP_Testsuite_AssertS(itm.description, "### TODO Expected string ###");
 };